pyproject.toml
[tool.poetry]
name = "pixelize"
version = "0.1.0"
description = ""
authors = ["Koutaro Mukai <mukai.k1011k@outlook.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.4.0"
numpy = "^1.24.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.0.0"
autopep8 = "^2.0.1"
opencv-python = "^4.7.0.68"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
